fix: run oclif-dev pack:deb as non root user
The only part of the pack:deb command that requires root are the chown and chgrp shell commands. I don't run oclif-dev as root, so the command fails at this point. By adding sudo, it asks for the password and only runs these commands with root privileges, which seems sensible to me.
The change will have no effect if already running with root privileges.