Use require to see if the project uses TypeScript
Instead of checking the package's devDependencies to determine if the
package uses TypeScript, try to `require('typescript')` and use success
as an indicator that the package uses TypeScript. This allows TypeScript
to be detected even if it is not listed in the project's dependencies
(such as when the package is in a yarn workspace and TypeScript is in
the dependencies of the workspace root).