build: only require .circleci/loadgpg in release job
The `OCLIF_DEV_DEB_SECRET_KEY` is only available in the "global"
circleCI context, which only the `release` job runs in. If an open
contributor submitting a pull request form a fork submits a pull
request, the build will fail due to this ENV var being present. Since we
don't release on pull requests, we can only run this in the `release`
job.